use uint64_t instead of uint32_t

+++ lab02/fib.c
@@ -1,14 +1,15 @@
 #include <stdio.h>
+#include <stdint.h>
 
-int fib(int n) {
+uint64_t fib(uint64_t n) {
   if (n == 1 || n == 2) return 1;
   return fib(n - 1) + fib(n - 2);
 }
 
-static const int N = 20;
+static const uint64_t N = 20;
 
 int main() {
-  int r = fib(N);
+  uint64_t r = fib(N);
   printf("fib(%d) = %d (%08x)\n", N, r, r);
   return 0;
 }
